feat: Update project name to Brokeflix and adjust related configurations for Windows support
