+++ lib/src/app_initializer.dart
@@ -3,6 +3,8 @@ import 'dart:developer';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
+import 'package:window_manager/window_manager.dart';
+import 'package:tray_manager/tray_manager.dart';
 
 import 'package:ytsmovies/src/app.dart';
 import 'package:ytsmovies/src/bloc/theme_bloc.dart';
@@ -22,11 +24,17 @@ class YTSAppInitializer extends StatefulWidget {
   State<YTSAppInitializer> createState() => _YTSAppInitializerState();
 }
 
-class _YTSAppInitializerState extends State<YTSAppInitializer> {
+class _YTSAppInitializerState extends State<YTSAppInitializer> with WindowListener, TrayListener {
   bool _isInitializing = true;
   String? _error;
   StreamSubscription<int>? _notificationSubscription;
-  DesktopWindowService? _desktopWindowService;
+  bool _desktopInitialized = false;
+
+  static const _trayShowKey = 'show_app';
+  static const _trayPauseKey = 'pause_all';
+  static const _trayResumeKey = 'resume_all';
+  static const _trayExitKey = 'exit_app';
+  static const _trayIconAsset = 'images/tray_icon.ico';
 
   @override
   void initState() {
@@ -46,9 +54,7 @@ class _YTSAppInitializerState extends State<YTSAppInitializer> {
 
       // Initialize desktop window + tray integration (no-op off desktop)
       if (isDesktop) {
-        _desktopWindowService =
-            DesktopWindowService(getIt<DownloadManagerBloc>());
-        await _desktopWindowService!.initialize();
+        await _initializeDesktopWindow();
       }
 
       setState(() {
@@ -61,6 +67,130 @@ class _YTSAppInitializerState extends State<YTSAppInitializer> {
         _isInitializing = false;
       });
     }
+  }
+
+  Future<void> _initializeDesktopWindow() async {
+    if (_desktopInitialized) return;
+    _desktopInitialized = true;
+
+    await windowManager.ensureInitialized();
+
+    const options = WindowOptions(
+      size: Size(1280, 800),
+      minimumSize: Size(900, 600),
+      center: true,
+      title: 'Brokeflix',
+      titleBarStyle: TitleBarStyle.normal,
+    );
+
+    await windowManager.waitUntilReadyToShow(options, () async {
+      await windowManager.setPreventClose(true);
+      await windowManager.show();
+      await windowManager.focus();
+    });
+
+    windowManager.addListener(this);
+    trayManager.addListener(this);
+
+    await _initTray();
+  }
+
+  Future<void> _initTray() async {
+    try {
+      await trayManager.setIcon(_trayIconAsset);
+      await trayManager.setToolTip('Brokeflix');
+      await _refreshTrayMenu();
+    } catch (e, s) {
+      log('Tray init failed: $e', error: e, stackTrace: s);
+    }
+  }
+
+  Future<void> _refreshTrayMenu() async {
+    final menu = Menu(
+      items: [
+        MenuItem(key: _trayShowKey, label: 'Open Brokeflix'),
+        MenuItem.separator(),
+        MenuItem(key: _trayPauseKey, label: 'Pause all downloads'),
+        MenuItem(key: _trayResumeKey, label: 'Resume paused downloads'),
+        MenuItem.separator(),
+        MenuItem(key: _trayExitKey, label: 'Exit'),
+      ],
+    );
+    await trayManager.setContextMenu(menu);
+  }
+
+  @override
+  void onWindowClose() async {
+    final prevent = await windowManager.isPreventClose();
+    if (!prevent) {
+      await windowManager.setPreventClose(true);
+    }
+    await windowManager.hide();
+    log('Window hidden to tray; downloads continue');
+  }
+
+  @override
+  void onTrayIconMouseDown() async {
+    await _showWindow();
+  }
+
+  @override
+  void onTrayIconRightMouseDown() async {
+    await trayManager.popUpContextMenu();
+  }
+
+  @override
+  void onTrayMenuItemClick(MenuItem menuItem) async {
+    switch (menuItem.key) {
+      case _trayShowKey:
+        await _showWindow();
+        break;
+      case _trayPauseKey:
+        _pauseAllActive();
+        break;
+      case _trayResumeKey:
+        _resumeAllPaused();
+        break;
+      case _trayExitKey:
+        await _exitApp();
+        break;
+    }
+  }
+
+  Future<void> _showWindow() async {
+    if (await windowManager.isMinimized()) {
+      await windowManager.restore();
+    }
+    await windowManager.show();
+    await windowManager.focus();
+  }
+
+  void _pauseAllActive() {
+    final downloadBloc = getIt<DownloadManagerBloc>();
+    for (final task in downloadBloc.state.downloads.values) {
+      if (task.canPause) {
+        downloadBloc.add(DownloadManagerPauseDownload(task.taskId));
+      }
+    }
+  }
+
+  void _resumeAllPaused() {
+    final downloadBloc = getIt<DownloadManagerBloc>();
+    for (final task in downloadBloc.state.downloads.values) {
+      if (task.status == DownloadStatus.paused) {
+        downloadBloc.add(DownloadManagerResumeDownload(task.taskId));
+      }
+    }
+  }
+
+  Future<void> _exitApp() async {
+    log('Tray exit: pausing active downloads before shutdown');
+    _pauseAllActive();
+    // Give libtorrent a moment to persist state.
+    await Future<void>.delayed(const Duration(milliseconds: 400));
+    await trayManager.destroy();
+    await windowManager.setPreventClose(false);
+    await windowManager.destroy();
   }
 
   void _handleNotificationTap(int taskId) {
@@ -92,7 +222,10 @@ class _YTSAppInitializerState extends State<YTSAppInitializer> {
   @override
   void dispose() {
     _notificationSubscription?.cancel();
-    _desktopWindowService?.dispose();
+    if (_desktopInitialized) {
+      windowManager.removeListener(this);
+      trayManager.removeListener(this);
+    }
     super.dispose();
   }
