+++ lib/src/pages/others.dart
@@ -12,7 +12,7 @@ class LatestMoviesPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return MoviesList(
       label: 'latest',
-      handler: (page) => context.read<MoviesClient>().getMovieList(page: page),
+      handler: (page, limit) => context.read<MoviesClient>().getMovieList(page: page, limit: limit),
       title: const Text('Latest Movies'),
       appBar: AppBar(
         title: Text(
@@ -32,8 +32,9 @@ class HD4KMoviesPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return MoviesList(
       label: '4k',
-      handler: (page) => context.read<MoviesClient>().getMovieList(
+      handler: (page, limit) => context.read<MoviesClient>().getMovieList(
             page: page,
+            limit: limit,
             quality: Quality.$2160,
           ),
       title: const Text('4K Movies'),
@@ -55,8 +56,9 @@ class RatedMoviesPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return MoviesList(
       label: 'rated',
-      handler: (page) => context.read<MoviesClient>().getMovieList(
+      handler: (page, limit) => context.read<MoviesClient>().getMovieList(
             page: page,
+            limit: limit,
             minimumRating: 5,
           ),
       title: const Text('Highly Rated Movies'),
