+++ lib/src/widgets/movies_paged_view.dart
@@ -64,7 +64,9 @@ class _MoviesPagedViewState extends State<MoviesPagedView> {
 
   void _fetchPage(int pageKey) async {
     try {
-      final response = await widget.handler(pageKey);
+      final crossAxisCount = context.posterGridColumns(targetItemWidth: 220);
+      final limit = (crossAxisCount * 3).clamp(10, 50);
+      final response = await widget.handler(pageKey, limit);
       final isLastPage = response.data.isLastPage;
       final movies = response.data.movies ?? [];
 
@@ -81,8 +83,10 @@ class _MoviesPagedViewState extends State<MoviesPagedView> {
   @override
   Widget build(BuildContext context) {
     final crossAxisCount = context.posterGridColumns(targetItemWidth: 220);
-    return CupertinoScrollbar(
+    return Scrollbar(
       controller: _scrollController,
+      thumbVisibility: true,
+      interactive: true,
       child: RefreshIndicator(
         onRefresh: () => SynchronousFuture(_pagingController.refresh()),
         child: PagedGridView<int, Movie>(
@@ -108,10 +112,12 @@ class _MoviesPagedViewState extends State<MoviesPagedView> {
                 movie: item,
               );
             },
-            firstPageProgressIndicatorBuilder: (_) => const MovieListShimmer(),
+            firstPageProgressIndicatorBuilder: (_) => MovieListShimmer(
+              count: (crossAxisCount * 3).clamp(10, 50),
+            ),
             newPageProgressIndicatorBuilder: (_) => Padding(
               padding: const EdgeInsets.only(top: 8.0),
-              child: const MovieListShimmer(count: 2),
+              child: MovieListShimmer(count: crossAxisCount),
             ),
             firstPageErrorIndicatorBuilder: _firstPageErrorIndicator,
             newPageErrorIndicatorBuilder: _newPageErrorIndicator,
