+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: ytsmovies
-description: YTS Movies — browse, search and download YTS movie torrents directly to your device.
+description: Brokeflix — browse, search and download movie torrents directly to your device.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `pub publish`. This is preferred for private packages.
